feat: 예외처리 - 2개 이상의 이름을 입력해야한다.
