+++ src/main/java/racingcar/view/InputView.java
@@ -3,12 +3,13 @@ package racingcar.view;
 import camp.nextstep.edu.missionutils.Console;
 import java.util.List;
 import racingcar.util.Util;
+import racingcar.util.Validator;
 
 public class InputView {
     public List<String> inputToNameCar() {
         System.out.println("경주할 자동차 이름을 입력하세요.(이름은 쉼표(,) 기준으로 구분)");
         List<String> inputs = Util.splitNameByComma(Console.readLine());
-        checkNameLength(inputs);
+        Validator.getInstance().validateCarName(inputs);
         return inputs;
     }
 
@@ -18,11 +19,5 @@ public class InputView {
         return input;
     }
 
-    private void checkNameLength(List<String> inputs) {
-        for (String input : inputs) {
-            if (input.length() > 5) {
-                throw new IllegalArgumentException("이름은 5자 이하만 입력할 수 있습니다.");
-            }
-        }
-    }
+
 }
